--- Client/src/calpolyData/interestAreas.json
@@ -1,18 +1,0 @@
-[
-  "Artificial Intelligence",
-  "Cybersecurity",
-  "Data Science",
-  "Software Engineering",
-  "Web Development",
-  "Mobile Development",
-  "Game Development",
-  "Machine Learning",
-  "Computer Vision",
-  "Natural Language Processing",
-  "Virtual Reality",
-  "Robotics",
-  "Blockchain",
-  "Quantum Computing",
-  "Augmented Reality",
-  "3D Printing"
-]
