+++ Client/src/components/layout/SectionPage/SectionPageLayout.tsx
@@ -2,11 +2,11 @@ import React from "react";
 
 // My components
 import SectionPageHeader from "./SectionPageHeader";
-import MobileHeader from "@/components/layout/MobileHeader";
+import MobileHeader from "@/components/layout/dynamicLayouts/MobileHeader";
 
 // Hooks
 import useIsNarrowScreen from "@/hooks/useIsNarrowScreen";
-import OuterSidebar from "../OuterIconSidebar";
+import OuterIconSidebar from "../dynamicLayouts/OuterIconSidebar";
 
 type SectionPageLayoutProps = {
   children: React.ReactNode;
@@ -18,7 +18,7 @@ const SectionPageLayout = ({ children }: SectionPageLayoutProps) => {
   return (
     <div className="flex flex-col h-full">
       <div className="flex">
-        {isNarrowScreen ? null : <OuterSidebar />}
+        {isNarrowScreen ? null : <OuterIconSidebar />}
         <div
           className={`bg-slate-900 text-white flex flex-col no-scroll w-full ${
             isNarrowScreen ? "ml-0" : "ml-2"
