+++ Client/src/constants/majors.ts
