+++ Client/src/constants/terms.ts
