--- Client/src/hooks/useCalpolyData.ts
@@ -1,6 +1,0 @@
-import interestAreas from "../calpolyData/interestAreas.json";
-const useCalpolyData = () => {
-  return { interestAreas };
-};
-
-export default useCalpolyData;
