+++ Client/src/pages/FlowchartPage.tsx
@@ -7,12 +7,12 @@ import {
 import { Flowchart, EmptyFlowchart } from "@/components/flowchart";
 import FlowchartLayout from "@/components/layout/FlowchartPage/FlowchartLayout";
 import { SidebarProvider } from "@/components/ui/sidebar";
-import OuterSidebar from "@/components/layout/OuterIconSidebar";
+import OuterIconSidebar from "@/components/layout/dynamicLayouts/OuterIconSidebar";
 import { useParams } from "react-router-dom";
 import { environment } from "@/helpers/getEnvironmentVars";
 import useIsNarrowScreen from "@/hooks/useIsNarrowScreen";
 
-const FlowChatPage = () => {
+const FlowChartPage = () => {
   const dispatch = useAppDispatch();
   const isNarrowScreen = useIsNarrowScreen();
   const { flowchartId } = useParams();
@@ -70,7 +70,7 @@ const FlowChatPage = () => {
   return (
     <>
       <div className="flex overflow-hidden no-scroll">
-        {isNarrowScreen ? null : <OuterSidebar />}
+        {isNarrowScreen ? null : <OuterIconSidebar />}
         <SidebarProvider className="dark:bg-slate-900">
           <FlowchartLayout>
             {flowchartData ? (
@@ -85,4 +85,4 @@ const FlowChatPage = () => {
   );
 };
 
-export default FlowChatPage;
+export default FlowChartPage;
