+++ Client/src/redux/index.ts
@@ -13,7 +13,6 @@ export { userReducer } from "./user/userSlice.ts";
 export { flowchartReducer } from "./flowchart/flowchartSlice.ts";
 export { flowSelectionReducer } from "./flowSelection/flowSelectionSlice.ts";
 export { sectionReducer } from "./section/sectionSlice.ts";
-export { panelLayoutReducer } from "./panelLayout/panelLayoutSlice.ts";
 export { sectionSelectionReducer } from "./sectionSelection/sectionSelectionSlice.ts";
 export { calendarReducer } from "./calendar/calendarSlice.ts";
 // Export all actions for easy dispatching in components
@@ -26,7 +25,6 @@ export * as userActions from "./user/userSlice.ts";
 export * as flowchartActions from "./flowchart/flowchartSlice.ts";
 export * as flowSelectionActions from "./flowSelection/flowSelectionSlice.ts";
 export * as sectionActions from "./section/sectionSlice.ts";
-export * as panelLayoutActions from "./panelLayout/panelLayoutSlice.ts";
 export * as sectionSelectionActions from "./sectionSelection/sectionSelectionSlice.ts";
 export * as calendarActions from "./calendar/calendarSlice.ts";
 // Custom hooks for using dispatch and selector with TypeScript
